+++ src/test/java/org/folio/rest/api/ItemStorageTest.java
@@ -113,6 +113,7 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
   private static final String INVALID_UUID = "invalid value";
   private static final String INVALID_TYPE_ERROR_MESSAGE = String.format("invalid input syntax for type uuid: \"%s\"",
     INVALID_UUID);
+  private static final String ORDER_FIELD = "order";
 
   private final ItemEventMessageChecks itemMessageChecks
     = new ItemEventMessageChecks(KAFKA_CONSUMER);
@@ -230,8 +231,9 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
     );
 
     JsonObject itemToCreate = new JsonObject();
-    itemToCreate.put("administrativeNotes", new JsonArray().add(adminNote));
     itemToCreate.put("id", id.toString());
+    itemToCreate.put(ORDER_FIELD, 100);
+    itemToCreate.put("administrativeNotes", new JsonArray().add(adminNote));
     itemToCreate.put("holdingsRecordId", holdingsRecordId.toString());
     itemToCreate.put("barcode", "565578437802");
     itemToCreate.put("status", new JsonObject().put("name", "Available"));
@@ -246,9 +248,7 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
     itemToCreate.put("itemLevelCallNumberSuffix", "allOwnComponentsCNS");
     itemToCreate.put("itemLevelCallNumberPrefix", "allOwnComponentsCNP");
     itemToCreate.put("itemLevelCallNumberTypeId", LC_CN_TYPE_ID);
-
     itemToCreate.put("statisticalCodeIds", List.of(statisticalCodeId));
-
     itemToCreate.put("inTransitDestinationServicePointId", inTransitServicePointId);
 
     setItemSequence(1);
@@ -256,61 +256,15 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
     var postResponse = saveItemAndExpectJson(itemToCreate);
 
     assertThat(postResponse.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
-
-    JsonObject itemFromPost = postResponse.getJson();
-
-    assertThat(itemFromPost.getString("id"), is(id.toString()));
-    assertThat(itemFromPost.getJsonArray("administrativeNotes").contains(adminNote), is(true));
-    assertThat(itemFromPost.getString("holdingsRecordId"), is(holdingsRecordId.toString()));
-    assertThat(itemFromPost.getString("barcode"), is("565578437802"));
-    assertThat(itemFromPost.getJsonObject("status").getString("name"),
-      is("Available"));
-    assertThat(itemFromPost.getString("displaySummary"), is(displaySummary));
-    assertThat(itemFromPost.getString("materialTypeId"),
-      is(journalMaterialTypeID));
-    assertThat(itemFromPost.getString("permanentLoanTypeId"),
-      is(canCirculateLoanTypeID));
-    assertThat(itemFromPost.getString("temporaryLocationId"),
-      is(ANNEX_LIBRARY_LOCATION_ID.toString()));
-    assertThat(itemFromPost.getString("inTransitDestinationServicePointId"),
-      is(inTransitServicePointId));
-    assertThat(itemFromPost.getString("hrid"), is("it00000000001"));
-    assertThat(itemFromPost.getString("copyNumber"), is("copy1"));
-    assertThat(itemFromPost.getString("effectiveShelvingOrder"),
-      is("PS 43623 R534 P37 42005 COP Y1 allOwnComponentsCNS"));
+    assertItem(postResponse.getJson(), id, adminNote, holdingsRecordId, displaySummary, inTransitServicePointId,
+      statisticalCodeId);
 
     Response getResponse = getById(id);
-
     assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_OK));
-
     JsonObject itemFromGet = getResponse.getJson();
 
-    assertThat(itemFromGet.getString("id"), is(id.toString()));
-    assertThat(itemFromGet.getJsonArray("administrativeNotes").contains(adminNote), is(true));
-    assertThat(itemFromGet.getString("holdingsRecordId"), is(holdingsRecordId.toString()));
-    assertThat(itemFromGet.getString("barcode"), is("565578437802"));
-    assertThat(itemFromGet.getJsonObject("status").getString("name"),
-      is("Available"));
-    assertThat(itemFromPost.getString("displaySummary"), is(displaySummary));
-    assertThat(itemFromGet.getString("materialTypeId"),
-      is(journalMaterialTypeID));
-    assertThat(itemFromGet.getString("permanentLoanTypeId"),
-      is(canCirculateLoanTypeID));
-    assertThat(itemFromGet.getString("temporaryLocationId"),
-      is(ANNEX_LIBRARY_LOCATION_ID.toString()));
-    assertThat(itemFromGet.getString("inTransitDestinationServicePointId"),
-      is(inTransitServicePointId));
-    assertThat(itemFromGet.getString("hrid"), is("it00000000001"));
-
-    List<String> tags = getTags(itemFromGet);
-
-    assertThat(tags.size(), is(1));
-    assertThat(tags, hasItem(TAG_VALUE));
-    assertThat(itemFromGet.getString("copyNumber"), is("copy1"));
-    assertThat(itemFromGet.getString("effectiveShelvingOrder"),
-      is("PS 43623 R534 P37 42005 COP Y1 allOwnComponentsCNS"));
-    assertThat(itemFromGet.getJsonArray("statisticalCodeIds"), hasItem(statisticalCodeId.toString()));
-
+    assertItem(itemFromGet, id, adminNote, holdingsRecordId, displaySummary, inTransitServicePointId,
+      statisticalCodeId);
     itemMessageChecks.createdMessagePublished(itemFromGet);
   }
 
@@ -341,6 +295,7 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
     assertThat(itemFromGet.getString("id"), is(id.toString()));
     assertThat(itemFromGet.getJsonObject("status").getString("name"),
       is("Available"));
+    assertThat(itemFromGet.getInteger(ORDER_FIELD), is(1));
 
     List<String> tags = getTags(itemFromGet);
 
@@ -348,6 +303,80 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
     assertThat(tags, hasItem(TAG_VALUE));
   }
 
+  @SneakyThrows
+  @Test
+  public void canCreateSeveralItemsThatWillCalculateDifferentOrder() {
+    var holdingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID);
+
+    var id1 = UUID.randomUUID();
+    var id2 = UUID.randomUUID();
+
+    var itemToCreate1 = minimalItem(id1, holdingsRecordId);
+    var itemToCreate2 = minimalItem(id2, holdingsRecordId);
+
+    var createCompleted1 = new CompletableFuture<Response>();
+    var createCompleted2 = new CompletableFuture<Response>();
+
+    getClient().post(itemsStorageUrl(""), itemToCreate1, TENANT_ID,
+      ResponseHandler.json(createCompleted1));
+    getClient().post(itemsStorageUrl(""), itemToCreate2, TENANT_ID,
+      ResponseHandler.json(createCompleted2));
+
+    var postResponse1 = createCompleted1.get(TIMEOUT, TimeUnit.SECONDS);
+    var postResponse2 = createCompleted2.get(TIMEOUT, TimeUnit.SECONDS);
+    assertThat(postResponse1.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+    assertThat(postResponse2.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+
+    var getResponse1 = getById(id1);
+    var getResponse2 = getById(id2);
+
+    assertThat(getResponse1.getJson().getInteger(ORDER_FIELD), anyOf(is(1), is(2)));
+    assertThat(getResponse2.getJson().getInteger(ORDER_FIELD), anyOf(is(1), is(2)));
+    assertThat(getResponse2.getJson().getInteger(ORDER_FIELD), not(is(getResponse1.getJson().getInteger(ORDER_FIELD))));
+  }
+
+  @SneakyThrows
+  @Test
+  public void canCreateSeveralItemsThatWillCalculateOrderWhenItDecreasesInNextRequest() {
+    var holdingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID);
+
+    var id1 = UUID.randomUUID();
+    var id2 = UUID.randomUUID();
+
+    var itemToCreate1 = minimalItem(id1, holdingsRecordId).put(ORDER_FIELD, 1000);
+    var itemToCreate2 = minimalItem(id2, holdingsRecordId).put(ORDER_FIELD, 10);
+
+    var createCompleted1 = new CompletableFuture<Response>();
+    var createCompleted2 = new CompletableFuture<Response>();
+
+    getClient().post(itemsStorageUrl(""), itemToCreate1, TENANT_ID,
+      ResponseHandler.json(createCompleted1));
+    getClient().post(itemsStorageUrl(""), itemToCreate2, TENANT_ID,
+      ResponseHandler.json(createCompleted2));
+
+    var postResponse1 = createCompleted1.get(TIMEOUT, TimeUnit.SECONDS);
+    var postResponse2 = createCompleted2.get(TIMEOUT, TimeUnit.SECONDS);
+    assertThat(postResponse1.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+    assertThat(postResponse2.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+
+    var id3 = UUID.randomUUID();
+    var itemToCreate3 = minimalItem(id3, holdingsRecordId);
+    var createCompleted3 = new CompletableFuture<Response>();
+    getClient().post(itemsStorageUrl(""), itemToCreate3, TENANT_ID,
+      ResponseHandler.json(createCompleted3));
+    var postResponse3 = createCompleted3.get(TIMEOUT, TimeUnit.SECONDS);
+    assertThat(postResponse3.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+
+    var getResponse1 = getById(id1);
+    var getResponse2 = getById(id2);
+    var getResponse3 = getById(id3);
+
+    assertThat(getResponse1.getJson().getInteger(ORDER_FIELD), is(1000));
+    assertThat(getResponse2.getJson().getInteger(ORDER_FIELD), is(10));
+    assertThat(getResponse3.getJson().getInteger(ORDER_FIELD), is(1001));
+  }
+
+  @SneakyThrows
   @Test
   public void cannotCreateAnItemWithInvalidStatisticalCodeIds() {
     UUID holdingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID);
@@ -1307,6 +1336,8 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
   @Test
   public void shouldCreateItemsWithPostSynchronousBatch() {
     JsonArray itemsArray = threeItems();
+    populateOrder(itemsArray);
+
     assertThat(postSynchronousBatch(itemsArray), statusCodeIs(HttpStatus.HTTP_CREATED));
     for (Object item : itemsArray) {
       assertExists((JsonObject) item);
@@ -1400,11 +1431,13 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
       .add(nod(existingItemId, holdingsRecordId))
       .add(smallAngryPlanet(holdingsRecordId))
       .add(interestingTimes(holdingsRecordId));
+    populateOrder(itemsArray1);
 
     final JsonArray itemsArray2 = new JsonArray()
       .add(nod(existingItemId, holdingsRecordId))
       .add(temeraire(holdingsRecordId))
       .add(uprooted(holdingsRecordId));
+    populateOrder(itemsArray2);
 
     final var firstResponse = postSynchronousBatch("?upsert=true", itemsArray1);
     final var existingItemBeforeUpdate = getById(existingItemId).getJson();
@@ -2866,6 +2899,34 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
+  public void shouldBatchCreateItems() {
+    final var items = threeItems();
+    var itemIds = JsonArrayHelper.toList(items)
+      .stream()
+      .map(item -> {
+        UUID itemId = UUID.randomUUID();
+        item.put("id", itemId.toString());
+        return itemId;
+      })
+      .toList();
+
+    final var response = itemsStorageSyncClient
+      .attemptToCreate(new JsonObject().put("items", items));
+
+    assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+
+    for (var id : itemIds) {
+      var getResponse = getById(id);
+      assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_OK));
+      var itemFromGet = getResponse.getJson();
+
+      assertThat(itemFromGet.getString("id"), is(id.toString()));
+      assertThat(itemFromGet.getString("hrid"), notNullValue());
+      assertThat(itemFromGet.getInteger(ORDER_FIELD), anyOf(is(1), is(2), is(3)));
+    }
+  }
+
+  @Test
   public void cannotCreateItemWithNonExistentHoldingsRecordId() {
     final UUID nonExistentHoldingsRecordId = randomUUID();
 
@@ -3369,6 +3430,16 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
     return item;
   }
 
+  private JsonObject minimalItem(UUID id, UUID holdingsRecordId) {
+    return new JsonObject()
+      .put("id", id.toString())
+      .put("status", new JsonObject().put("name", "Available"))
+      .put("holdingsRecordId", holdingsRecordId.toString())
+      .put("materialTypeId", journalMaterialTypeID)
+      .put("permanentLoanTypeId", canCirculateLoanTypeID)
+      .put("tags", new JsonObject().put("tagList", new JsonArray().add(TAG_VALUE)));
+  }
+
   private Response postSynchronousBatchWithExistingId(String subPath) {
     JsonArray itemsArray1 = threeItems();
     JsonArray itemsArray2 = threeItems();
@@ -3401,6 +3472,38 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
       .put("holdingsRecordId", holdingsRecordId.toString())
       .put("materialTypeId", journalMaterialTypeID)
       .put("permanentLoanTypeId", canCirculateLoanTypeID);
+  }
+
+  private void populateOrder(JsonArray itemsArray2) {
+    for (int i = 0; i < itemsArray2.size(); i++) {
+      JsonObject item = itemsArray2.getJsonObject(i);
+      item.put(ORDER_FIELD, i);
+    }
+  }
+
+  private void assertItem(JsonObject itemFromGet, UUID id, String adminNote, UUID holdingsRecordId,
+                          String displaySummary, String inTransitServicePointId, UUID statisticalCodeId) {
+    assertThat(itemFromGet.getString("id"), is(id.toString()));
+    assertThat(itemFromGet.getInteger(ORDER_FIELD), is(100));
+    assertThat(itemFromGet.getJsonArray("administrativeNotes").contains(adminNote), is(true));
+    assertThat(itemFromGet.getString("holdingsRecordId"), is(holdingsRecordId.toString()));
+    assertThat(itemFromGet.getString("barcode"), is("565578437802"));
+    assertThat(itemFromGet.getJsonObject("status").getString("name"), is("Available"));
+    assertThat(itemFromGet.getString("displaySummary"), is(displaySummary));
+    assertThat(itemFromGet.getString("materialTypeId"), is(journalMaterialTypeID));
+    assertThat(itemFromGet.getString("permanentLoanTypeId"), is(canCirculateLoanTypeID));
+    assertThat(itemFromGet.getString("temporaryLocationId"), is(ANNEX_LIBRARY_LOCATION_ID.toString()));
+    assertThat(itemFromGet.getString("inTransitDestinationServicePointId"), is(inTransitServicePointId));
+    assertThat(itemFromGet.getString("hrid"), is("it00000000001"));
+
+    List<String> tags = getTags(itemFromGet);
+
+    assertThat(tags.size(), is(1));
+    assertThat(tags, hasItem(TAG_VALUE));
+    assertThat(itemFromGet.getString("copyNumber"), is("copy1"));
+    assertThat(itemFromGet.getString("effectiveShelvingOrder"),
+      is("PS 43623 R534 P37 42005 COP Y1 allOwnComponentsCNS"));
+    assertThat(itemFromGet.getJsonArray("statisticalCodeIds"), hasItem(statisticalCodeId.toString()));
   }
 
   private static JsonObject createItemRequest(